Implement Send button for submissions (issue 54)
When the Send button of the Submission details page is clicked, a POST
is now made to the URL and port specified in submission.js.  The code
recognizes the OK/fail code that is returned by the CGI server from the
command it has executed (for PASS, this is an FTP operation).

Note that if the initial version of the CGI server is used, the command
will be executed but the return value will not be recognized.  This is
because that version of the server does not allow responses to be sent
for POST requests.  A pull request will be made for an updated version
of the CGI script that is allowed to reply to all requestors, among
other changes.  This may not be the most desirable behavior in the long
term, nor the only possible solution to this problem, but it is needed
at this time.