Make rootURL available to all pages
Closes issue #44
Added rootURL environment variable as extension to default controller so
that it could be used to form the logo path on any page. Previously, the
logo was only working on paths with one subfolder because src="logo.png"
was sufficient to find the image in the root directory. Adding
sub-folders as in `/grants/1` caused the relative path to fail
(../logo.png would have been required).