Disable local storage for submission, fix bugs
Because of some issues with the model, loading the submission from local storage in the route currently breaks the page. This commit comments out that behavior. This commit also updates any references to "model" to refer instead to "model.newSubmission", and updates the submission nav to use the right variables.