Add USAID link support (#621)
#614

Add support for USAID grants in a submission. Previously, these grants were basically ignored in the submission workflow, thus in the submission details page as well. This update will make USAID and other web-link grants behave similarly to Dept of Edu grants: submission will be halted at the review step. Final submission will only be possible after the user clicks a link to the open the web-linked repository (with the goal of having the user manually add a submission).