Update images and point to assets without ACLs
The ACL update is a workaround based off of the 6/19 assets image
(containing the 5 and 7 year NIHMS and COEUS loads).  It contains the raw
data from that dump, plus coarse grained ACLs only.

Additionally, the authz updater has been removed until an updated version
is published

This should allow Ember development to cuntinue without 403s untul the
authz updater image is published, and a new assets image is published.