Change way items looked up in externalRepoMap
This is an attempt to fix issue #714. Interestingly there was a line of code commented out that should fix it! In production, repository IDs have periods in them, in `pass-docker` they do not. I believe the way the code was previously was causing ember to interpret the periods as a series of gets `get("pass").get("jhu").get("edu")` etc.