Details with different expression of timeout in production
In production, timeout results in an error being thrown by whoami due to the fact shib is on a different domain name and js can't reach it. The error is caught by the default handler instead so the default error is shown instead of the timeout error. This adds a catch so that the timeout error is shown.