Refactored the submission workflow to be on 8 routes instead of 1 (#861)
Having the entire workflow on a single route has meant each tab does not have it's own route/controller layer for loading and processing data specific to the tab. As the application becomes more complicated and data intensive, the "new" controller shared by 7 tabs is getting very large. This change splits the single route into 8 - the parent "new" route and one for each tab, so that each can have it's own model and controller objects specific to the tab. This change is purely about restructuring, further refactoring will aim to move controller information relevant to a single tab into the appropriate controller. To simplify the change, a new "workflow" service has been created to hold properties that used to exist in the workflow-wrapper. When there was one route the workflow-wrapper could be used to pass data into components but with 7 tab routes, its information changes for each so is no longer viable. The workflow service is a bunch of setters and getters that replace the job of the workflow-wrapper. We can either continue with this approach, or add properties to the parent controller - this will be addressed (or not if it is viable as is) in a future change. Note that tests have not been updated to reflect the new model since refactoring to take place over the coming weeks will require them to be rewritten.