Change "on behalf of" radio button to use action
Previously, because of some issues with how the radio-button object works, an observer was used on the radio button. This switches to use an html input radio button so that an action on toggle can be used. Some code added to preserve the relationship between `newSubmission.hasNewProxy` and the button.