Formalized isProxySubmission as computed property on Submission model
Originally, the workflow used a `hasProxy` property that lived on the new route. Since all properties necessary to calculate this exist on the model, this change ensures the workflow uses it.
While getting this to work in the workflow-basics component, I also removed a chunk of redundant code for changing the submitter.