Removed code for didAgree and didNotAgree properties in new controller
It seems that previously, if JScholarship was added as a repository and then the agreement wasn't signed, a notice would be displayed on the review page and the repository removed on submit. Now, however, the repository is removed during the workflow if the person does not agree to the agreement. This change removes unused residual old code.